Clean up ErrorResponse, use async await
