+++ packages/neoBackend/package.json
@@ -21,6 +21,7 @@
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
     "express": "^4.17.1",
+    "express-async-handler": "^1.1.4",
     "jsonwebtoken": "^8.5.1",
     "node-fetch": "^2.6.0",
     "ws": "^7.0.1"
